Support wildcard certificates with domain aliases
Wildcard certificates that begin with an asterisk cannot
be processed by the f5acmehandler.sh wrapper script due
to file globbing issues with the shell. Using an alias
instead (i.e. star_example_com for *.example.com) fixes
the issue.  Allows the use of the --domain flag to support
SANs along with wildcard certificates.

Also update the README.md to include documentation on how
to use the alias feature with a wildcard certificate.